requirements.txt
irc
PyMySQL
PyYAML
requests
sqlalchemy
phpserialize